+++ src/definitions/types/Datalist.ts
@@ -1,8 +1,15 @@
 export interface BillDataAPI {
   id: string;
+  bill_name: string;
+  owner: string;
+  published_date: string;
+  bill_status: boolean;
+  bill_image: string;
+}
+
+export interface BankData {
+  id: string;
   name: string;
-  owner_id: string;
-  publish_date: string;
-  status: boolean;
-  image: string;
+  logo: string;
+  format: string;
 }