+++ src/mocks/BankData.tsx
@@ -1,0 +1,46 @@
+export const data = {
+  banks: [
+    {
+      id: '19384789347',
+      name: 'พร้อมเพย์',
+      logo : 'https://storage.googleapis.com/billme-static/billme/bank-logo/logo-bank-promptpay.png',
+      format: 'xxx-x-xxxxx-x'
+    },
+    {
+      id: '19384789347',
+      name: 'เกียรตินาคิน',
+      logo : 'https://storage.googleapis.com/billme-static/billme/bank-logo/logo-bank-kiatnakin.png',
+      format: 'xxx-x-xxxxx-x'
+    },
+    {
+      id: '19384789347',
+      name: 'ไทยพาณิชย์',
+      logo : 'https://storage.googleapis.com/billme-static/billme/bank-logo/logo-bank-scb.png',
+      format: 'xxx-x-xxxxx-x'
+    },
+    {
+      id: '19384789347',
+      name: 'กสิกรไทย',
+      logo : 'https://storage.googleapis.com/billme-static/billme/bank-logo/logo-bank-kbank.png',
+      format: 'xxx-x-xxxxx-x'
+    },
+    {
+      id: '19384789347',
+      name: 'กรุงเทพ',
+      logo : 'https://storage.googleapis.com/billme-static/billme/bank-logo/logo-bank-bkk.png',
+      format: 'xxx-x-xxxxx-x'
+    },
+    {
+      id: '19384789347',
+      name: 'ออมสิน',
+      logo : 'https://storage.googleapis.com/billme-static/billme/bank-logo/logo-bank-gsb.png',
+      format: 'xxx-x-xxxxx-x'
+    },
+    {
+      id: '19384789347',
+      name: 'กรุงไทย',
+      logo : 'https://storage.googleapis.com/billme-static/billme/bank-logo/logo-bank-ktb.png',
+      format: 'xxx-x-xxxxx-x'
+    }
+  ]
+};