+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [V0.0.13] - 2019-06-25
+## Updated
+- Added Dropdown Component.
+
 ## [V0.0.12] - 2019-06-25
 ## Updated
 - Added Friend list template.